treat array and object the same (more general)
Due to the fact that there are multiple terms, at different depth levels, that can be objects or arrays, we must decide on a common behaviour at the first depth level. The idea is that whenever the property is a more complex object/array, we assign it a new class, and then decide how to model that.